add delete/edit/search api

+++ src/main/java/org/example/controllers/CategoryController.java
@@ -2,13 +2,16 @@ package org.example.controllers;
 
 import lombok.AllArgsConstructor;
 import org.example.dto.category.CategoryCreateDTO;
+import org.example.dto.category.CategoryEditDTO;
 import org.example.dto.category.CategoryItemDTO;
 import org.example.entities.CategoryEntity;
 import org.example.mapper.CategoryMapper;
 import org.example.repositories.CategoryRepository;
 import org.example.storage.FileSaveFormat;
 import org.example.storage.StorageService;
+import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -43,28 +46,63 @@ public class CategoryController {
         }
     }
 
-    @PostMapping(path = "/edit/{id}", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<CategoryItemDTO> edit(@ModelAttribute CategoryCreateDTO dto, @PathVariable("id") Integer id) {
-        try {
-            Optional<CategoryEntity> tmp = categoryRepository.findById(id);
-            CategoryEntity entity = categoryMapper.categoryCreateDTO(dto);
-            entity.setId(tmp.get().getId());
-            String image = storageService.saveImage(dto.getImage(), FileSaveFormat.WEBP);
-            entity.setImage(image);
-            categoryRepository.save(entity);
-            return new ResponseEntity<>(categoryMapper.categoryItemDTO(entity), HttpStatus.CREATED);
-        } catch (Exception ex) {
-            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+    @GetMapping("/{id}")
+    public ResponseEntity<CategoryItemDTO> getById(@PathVariable int id) {
+        var entity = categoryRepository.findById(id).orElse(null);
+        if (entity == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
+        var result =  categoryMapper.categoryItemDTO(entity);
+        return new ResponseEntity<>(result, HttpStatus.OK);
+    }
+
+    @PutMapping(path="/edit/{id}", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    public ResponseEntity<CategoryItemDTO> edit(@ModelAttribute CategoryEditDTO model) {
+        var old = categoryRepository.findById(model.getId()).orElse(null);
+        if (old == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        var entity = categoryMapper.categoryEditDto(model);
+        if(model.getFile()==null) {
+            entity.setImage(old.getImage());
+        }
+        else {
+            try {
+                storageService.deleteImage(old.getImage());
+                String fileName = storageService.saveImage(model.getFile(), FileSaveFormat.WEBP);
+                entity.setImage(fileName);
+            }
+            catch (Exception exception) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+        }
+        entity.setCreationTime(old.getCreationTime());
+        categoryRepository.save(entity);
+        var result = categoryMapper.categoryItemDTO(entity);
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<CategoryItemDTO> delete(@PathVariable("id") Integer id) {
-        try {
-            categoryRepository.deleteById(id);
-            return new ResponseEntity<>(HttpStatus.OK);
-        } catch (Exception ex) {
-            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+    public ResponseEntity<Void> delete(@PathVariable int id) {
+        var entity = categoryRepository.findById(id).orElse(null);
+        if (entity == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
+        try {
+            storageService.deleteImage(entity.getImage());
+            categoryRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        catch (Exception exception) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+    }
+
+    @GetMapping("/search")
+    public ResponseEntity<Page<CategoryItemDTO>> searchByName(@RequestParam(required = false) String name,
+                                                              Pageable pageable) {
+        Page<CategoryEntity> categories = categoryRepository.findByNameContainingIgnoreCase(name, pageable);
+        Page<CategoryItemDTO> result = categories.map(categoryMapper::categoryItemDTO);
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 }
